+++ src/infiniop/devices/cpu/common_cpu.cc
@@ -60,8 +60,8 @@ uint16_t f32_to_f16(float val) {
 }
 
 size_t indexToReducedOffset(size_t flat_index, size_t ndim,
-                            int64_t const *broadcasted_strides,
-                            int64_t const *target_strides) {
+                            ptrdiff_t const *broadcasted_strides,
+                            ptrdiff_t const *target_strides) {
     size_t res = 0;
     for (size_t i = 0; i < ndim; ++i) {
         res += flat_index / broadcasted_strides[i] * target_strides[i];
@@ -71,7 +71,7 @@ size_t indexToReducedOffset(size_t flat_index, size_t ndim,
 }
 
 size_t indexToOffset(size_t flat_index, size_t ndim, size_t const *shape,
-                     int64_t const *strides) {
+                     ptrdiff_t const *strides) {
     size_t res = 0;
     for (size_t i = ndim; i-- >= 0;) {
         res += (flat_index % shape[i]) * strides[i];
@@ -81,7 +81,7 @@ size_t indexToOffset(size_t flat_index, size_t ndim, size_t const *shape,
 }
 
 size_t getPaddedSize(size_t ndim, size_t *shape, size_t const *pads) {
-    uint64_t total_size = 1;
+    size_t total_size = 1;
     for (size_t i = 0; i < ndim; ++i) {
         total_size *= shape[i] + (i < 2 ? 0 : 2 * pads[i - 2]);
     }
